move db data initialization to tools suite api service
